@@ -6,7 +6,7 @@
     "homepage": "https://github.com/wp-media/wp-rocket-cli",
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4"
     },
     "authors": [{
         "name": "WP Rocket Team",
